Fix x64 compilation and types, add ref counted pointer type for 32 and 64-bit. Add rpmalloc.

+++ Simulation/src/ste/engine/graphics_interface/command_buffer/command/cmd_begin_render_pass.hpp
@@ -45,7 +45,7 @@ private:
 		info.renderPass = render_pass.get();
 		info.renderArea = render_area;
 		info.framebuffer = framebuffer.get();
-		info.clearValueCount = clear_values.size();
+		info.clearValueCount = static_cast<std::uint32_t>(clear_values.size());
 		info.pClearValues = clear_values.data();
 
 		vkCmdBeginRenderPass(command_buffer, &info, VK_SUBPASS_CONTENTS_INLINE);
