+++ Simulation/src/ste/engine/graphics_interface/command_buffer/command/cmd_copy_buffer_to_image.hpp
@@ -47,7 +47,7 @@ private:
 	void operator()(const command_buffer &command_buffer, command_recorder &) const override final {
 		vkCmdCopyBufferToImage(command_buffer, src_buffer.get().get_buffer_handle(),
 							   dst_image.get().get_image_handle(), static_cast<VkImageLayout>(layout),
-							   ranges.size(), ranges.data());
+							   static_cast<std::uint32_t>(ranges.size()), ranges.data());
 	}
 };
