+++ Simulation/src/ste/engine/graphics_interface/command_buffer/command/cmd_copy_image.hpp
@@ -41,8 +41,10 @@ public:
 private:
 	void operator()(const command_buffer &command_buffer, command_recorder &) const override final {
 		vkCmdCopyImage(command_buffer, src_image.get().get_image_handle(), static_cast<VkImageLayout>(src_image_layout),
-					   dst_image.get().get_image_handle(), static_cast<VkImageLayout>(dst_image_layout),
-					   ranges.size(), ranges.data());
+					   dst_image.get().get_image_handle(), 
+					   static_cast<VkImageLayout>(dst_image_layout),
+					   static_cast<std::uint32_t>(ranges.size()), 
+					   ranges.data());
 	}
 };
